requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-community
langchain-huggingface
langgraph
faiss-cpu
sentence-transformers
pypdf
scikit-learn
numpy
pandas
pytesseract
requests
jinja2
langchain-google-genai
pytest
